requirements.txt
tk-tools
nltk
pandas
tensorflow
streamlit
chatterbot==1.0.2
chatterbot-corpus
langchain
openai
tiktoken
pinecone-client
nest_asyncio